tests/test_select_main_page.py
import time
import allure

from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from pages.main_page import Main_page
from selenium.webdriver.chrome.service import Service


@allure.description("Test select main page")
def test_select_main_page():
    options = Options()
    options.add_argument("--user-agent=Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/127.0.0.0 Safari/537.36")
    options.add_experimental_option("excludeSwitches", ["enable-automation"])
    options.add_experimental_option("useAutomationExtension", False)
    options.add_argument("--disable-blink-features=AutomationControlled")
    options.add_argument("--no-sandbox")
    options.add_argument("--disable-dev-shm-usage")
    options.add_argument("--disable-gpu")
    options.add_argument("--headless=new")
    options.add_argument("--window-size=1920,1080")
    options.binary_location = "/usr/bin/chromium"

    service = Service(executable_path="/usr/bin/chromedriver")
    driver = webdriver.Chrome(service=service, options=options)
    print("Start test")

    mp = Main_page(driver)
    mp.select_main_page()

    time.sleep(10)
    driver.quit()